style: css 속성 및 코드 정리
